+++ bw.workflow.registerconvictednotpresent/src/main/java/sa/gov/nic/bio/bw/workflow/registerconvictednotpresent/RegisterConvictedReportNotPresentWorkflow.java
@@ -8,6 +8,7 @@ import sa.gov.nic.bio.bw.core.workflow.Signal;
 import sa.gov.nic.bio.bw.core.workflow.WithLookups;
 import sa.gov.nic.bio.bw.core.workflow.WizardWorkflowBase;
 import sa.gov.nic.bio.bw.workflow.commons.beans.ConvictedReport;
+import sa.gov.nic.bio.bw.workflow.commons.beans.DisCriminalReport;
 import sa.gov.nic.bio.bw.workflow.commons.beans.NormalizedPersonInfo;
 import sa.gov.nic.bio.bw.workflow.commons.beans.PersonInfo;
 import sa.gov.nic.bio.bw.workflow.commons.controllers.InquiryByFingerprintsPaneFxController;
@@ -19,8 +20,10 @@ import sa.gov.nic.bio.bw.workflow.commons.lookups.DocumentTypesLookup;
 import sa.gov.nic.bio.bw.workflow.commons.lookups.PersonTypesLookup;
 import sa.gov.nic.bio.bw.workflow.commons.tasks.ConvertFingerprintBase64ImagesToWsqWorkflowTask;
 import sa.gov.nic.bio.bw.workflow.commons.tasks.ConvertWsqFingerprintsToSegmentedFingerprintBase64ImagesWorkflowTask;
-import sa.gov.nic.bio.bw.workflow.commons.tasks.ConvictedReportInquiryWorkflowTask;
+import sa.gov.nic.bio.bw.workflow.commons.tasks.ConvictedReportInquiryByGeneralFileNumberWorkflowTask;
+import sa.gov.nic.bio.bw.workflow.commons.tasks.ConvictedReportInquiryFromDisWorkflowTask;
 import sa.gov.nic.bio.bw.workflow.commons.tasks.ConvictedReportToPersonInfoConverter;
+import sa.gov.nic.bio.bw.workflow.commons.tasks.DisCriminalReportToPersonInfoConverter;
 import sa.gov.nic.bio.bw.workflow.commons.tasks.FetchingFingerprintsWorkflowTask;
 import sa.gov.nic.bio.bw.workflow.commons.tasks.FetchingMissingFingerprintsWorkflowTask;
 import sa.gov.nic.bio.bw.workflow.commons.tasks.FingerprintInquiryStatusCheckerWorkflowTask;
@@ -71,8 +74,10 @@ import java.util.stream.Collectors;
 		@Step(iconId = "file_pdf_alt", title = "wizard.showReport")})
 public class RegisterConvictedReportNotPresentWorkflow extends WizardWorkflowBase
 {
+	private static final String FIELD_CIVIL_HIT = "CIVIL_HIT";
 	private static final String FIELD_CIVIL_PERSON_INFO_MAP = "CIVIL_PERSON_INFO_MAP";
-	private static final String FIELD_CRIMINAL_PERSON_INFO_MAP = "CRIMINAL_PERSON_INFO_MAP";
+	private static final String FIELD_OLD_CRIMINAL_PERSON_INFO_MAP = "OLD_CRIMINAL_PERSON_INFO_MAP";
+	private static final String FIELD_NEW_CRIMINAL_PERSON_INFO_MAP = "NEW_CRIMINAL_PERSON_INFO_MAP";
 	
 	@Override
 	public ResourceBundle getStringsResourceBundle(Locale locale)
@@ -265,6 +270,7 @@ public class RegisterConvictedReportNotPresentWorkflow extends WizardWorkflowBas
 					                                    "criminalBiometricsId");
 					if(civilBiometricsId != null)
 					{
+						setData(getClass(), FIELD_CIVIL_HIT, Boolean.TRUE);
 						List<Long> civilPersonIds = getData(FingerprintInquiryStatusCheckerWorkflowTask.class,
 						                                    "civilPersonIds");
 						if(!civilPersonIds.isEmpty())
@@ -288,20 +294,35 @@ public class RegisterConvictedReportNotPresentWorkflow extends WizardWorkflowBas
 					
 					if(criminalBiometricsId != null)
 					{
-						setData(ConvictedReportInquiryWorkflowTask.class, "criminalBiometricsId",
-						        criminalBiometricsId);
-						setData(ConvictedReportInquiryWorkflowTask.class,
+						setData(ConvictedReportInquiryFromDisWorkflowTask.class,
+						        "criminalBiometricsId", criminalBiometricsId);
+						setData(ConvictedReportInquiryFromDisWorkflowTask.class,
 						        "returnNullResultInCaseNotFound", Boolean.TRUE);
-						executeWorkflowTask(ConvictedReportInquiryWorkflowTask.class);
-						List<ConvictedReport> convictedReports = getData(ConvictedReportInquiryWorkflowTask.class,
-						                                                 "convictedReports");
-						ConvictedReportToPersonInfoConverter converter = new ConvictedReportToPersonInfoConverter();
-						Map<Long, PersonInfo> criminalPersonInfoMap;
-						if(convictedReports != null) criminalPersonInfoMap = convictedReports.stream().collect(
-								Collectors.toMap(ConvictedReport::getReportNumber, converter::convert,
+						executeWorkflowTask(ConvictedReportInquiryFromDisWorkflowTask.class);
+						List<DisCriminalReport> disCriminalReports =
+													getData(ConvictedReportInquiryFromDisWorkflowTask.class,
+													        "disCriminalReports");
+						DisCriminalReportToPersonInfoConverter converter = new DisCriminalReportToPersonInfoConverter();
+						Map<Integer, PersonInfo> oldCriminalPersonInfoMap;
+						if(disCriminalReports != null) oldCriminalPersonInfoMap = disCriminalReports.stream().collect(
+								Collectors.toMap(DisCriminalReport::getSequenceNumber, converter::convert,
 								                 (k1, k2) -> k1, LinkedHashMap::new));
-						else criminalPersonInfoMap = new LinkedHashMap<>();
-						setData(getClass(), FIELD_CRIMINAL_PERSON_INFO_MAP, criminalPersonInfoMap);
+						else oldCriminalPersonInfoMap = new LinkedHashMap<>();
+						setData(getClass(), FIELD_OLD_CRIMINAL_PERSON_INFO_MAP, oldCriminalPersonInfoMap);
+						
+						setData(ConvictedReportInquiryByGeneralFileNumberWorkflowTask.class,
+						        "criminalBiometricsId", criminalBiometricsId);
+						executeWorkflowTask(ConvictedReportInquiryByGeneralFileNumberWorkflowTask.class);
+						List<ConvictedReport> convictedReports =
+								getData(ConvictedReportInquiryByGeneralFileNumberWorkflowTask.class,
+								        "convictedReports");
+						ConvictedReportToPersonInfoConverter converter2 = new ConvictedReportToPersonInfoConverter();
+						Map<Long, PersonInfo> newCriminalPersonInfoMap;
+						if(convictedReports != null) newCriminalPersonInfoMap = convictedReports.stream().collect(
+								Collectors.toMap(ConvictedReport::getReportNumber, converter2::convert,
+								                 (k1, k2) -> k1, LinkedHashMap::new));
+						else newCriminalPersonInfoMap = new LinkedHashMap<>();
+						setData(getClass(), FIELD_NEW_CRIMINAL_PERSON_INFO_MAP, newCriminalPersonInfoMap);
 					}
 				}
 				
@@ -311,8 +332,10 @@ public class RegisterConvictedReportNotPresentWorkflow extends WizardWorkflowBas
 			{
 				passData(getClass(), FIELD_CIVIL_PERSON_INFO_MAP, InquiryByFingerprintsResultPaneFxController.class,
 				         "civilPersonInfoMap");
-				passData(getClass(), FIELD_CRIMINAL_PERSON_INFO_MAP, InquiryByFingerprintsResultPaneFxController.class,
-				         "criminalPersonInfoMap");
+				passData(getClass(), FIELD_OLD_CRIMINAL_PERSON_INFO_MAP,
+				         InquiryByFingerprintsResultPaneFxController.class, "oldCriminalPersonInfoMap");
+				passData(getClass(), FIELD_NEW_CRIMINAL_PERSON_INFO_MAP,
+				         InquiryByFingerprintsResultPaneFxController.class, "newCriminalPersonInfoMap");
 				setData(InquiryByFingerprintsResultPaneFxController.class, "hideRegisterUnknownButton",
 				        Boolean.TRUE);
 				passData(FingerprintInquiryStatusCheckerWorkflowTask.class,
@@ -324,6 +347,7 @@ public class RegisterConvictedReportNotPresentWorkflow extends WizardWorkflowBas
 			}
 			case 6:
 			{
+				passData(getClass(), FIELD_CIVIL_HIT, UpdatePersonInfoPaneFxController.class, "civilHit");
 				passData(InquiryByFingerprintsResultPaneFxController.class, UpdatePersonInfoPaneFxController.class,
 				         "normalizedPersonInfo");
 				renderUiAndWaitForUserInput(UpdatePersonInfoPaneFxController.class);