+++ RELEASE-NOTES.md
@@ -18,6 +18,7 @@ VERSION SCHEMA: vYYYY.MM.#SEQ
 - FIX: Prevent blocking the application navigation in case one the of the lookups takes long time.
 - ENH: In case of invalid state in workflows, navigate the user to the home page and set error label.
 - ENH: Upgrade ScenicView.
+- CHN: Update the fingerprint inquiry results to include records from DIS.
 
 ---